+++ src/routes/(v2)/v2/Navigation.svelte
@@ -76,7 +76,8 @@
 		'transport',
 		'bigint',
 		'issue-455',
-		'vine'
+		'vine',
+		'timeproxy'
 	].sort();
 </script>