+++ src/routes/(v2)/v2/timeproxy/schema.ts
@@ -1,0 +1,6 @@
+import { z } from 'zod';
+
+export const schema = z.object({
+	time: z.date(),
+	datetime: z.date()
+});